+++ src/entities/marking/ui/index.ts
@@ -1,1 +1,2 @@
 export * from "./markingList";
+export * from "./dashboard";
