+++ src/entities/profile/ui/index.ts
@@ -1,3 +1,1 @@
-export * from "./TemporaryMarkingBar";
-export * from "./GalleryGrid";
 export * from "./profile";
